Add get() to htcondor.param if missing
Old versions of condor-python do not have htcondor.param.get(); monkey-patch
htcondor.param if that is the case. This makes the code less repetitive than
using try/except blocks.

Also, "'foo' in htcondor.param" doesn't work in old versions of condor-python
so replace uses of that with htcondor.param.get()