Fix up default security settings.
This locks down the OWNER and NEGOTIATOR authorization levels; while
these changes do not do anything (as the daemons we run don't use
these in current versions of HTCondor), this should decrease the
possibility of security issues if these authorization levels are
used in a future version of HTCondor.

Additionally, this fixes ALLOW_ADMINISTRATOR to use FULL_HOSTNAME
instead of HOSTNAME, making it more likely that the entry matches
the CN of the hostcert.  This should fix problems with doing a
condor_reconfig on hosts.