Remove TotalSubmitProcs attribute in routed jobs
TotalSubmitProcs is used in 8.5.7 for the for job totals in the new
condor_q formatting. If a user queues multiple jobs in their submit
file (resulting in multiple ProcIDs), the job router creates unique
ClusterIDs in the target SchedD for each ProcID. Each of these routed
jobs retains the TotalSubmitProcs from the original jobs and condor_q
reports erroneously large job totals.