Send first keepalive after dropping privs (SOFTWARE-3486)
CherryPy drops privileges during the engine.start() call so the now
removed heartbeat was supposed to send heartbeats to the CE master
daemon with EUID 'root'. The original intent was to create the initial
security session with GSI credentials but that caused issues when the
security session expired after 24 hours (SOFTWARE-2820) so we use FS
auth for the CE View.

Due to an HTCondor
bug (https://htcondor-wiki.cs.wisc.edu/index.cgi/tktview?tn=6817),
Python procceses using the htcondor module would change their EUID to
the 'condor' user, which is part of the ALLOW_DAEMON list, so this all worked.
However, when the bug was fixed, the EUID remained as the 'root' user,
which is not part of the ALLOW_DAEMON list so the initial heartbeat
would fail. As an added bonus, htcondor.send_alive() raises a
RuntimeError on authZ failure so the CE View would exit 1.