Fix handling of unmapped GSI users (SOTWARE-4039)
For reference, HTCondor authentication with a mapfile looks like this:

1. Server and client negotiate an auth method
2. Server uses all the lines in the mapfile corresponding to the
chosen auth method. If the chosen auth method doesn't have any lines
in the mapfile, I believe there are default mapped IDs.
3. Staring with the top line, the server attempts to match the
client's credentials to the second field of each line

So in our case, every single GSI-negotiated mapping was matching to
the first line in the mapfile since the regex is `(.*)`. If the GSI
callout (i.e., `GSS_ASSIST_GRIDMAP`) fails, that results in the mapped
identity of `gsi@unmapped` and the mapfile is not consulted
again. Therefore, we can get rid of the second GSI line in the mapfile
and add `gsi@unmapped` to `UNMAPPED_USERS` to allow unmapped GSI hosts
to report to the central collector.