avoid counting unknown statuses (SOFTWARE-4183)
the previous behavior was an if/elif/elif for the different statuses
we care about, and did not have a final "else", meaning that other
statuses were not recorded.

with the recent

  status = status_map.get(job.get("JobStatus"))

thing, status is None if it's not a status we care about, so there is
a catch-all results[None] bucket for other statuses we don't care about.

We avoid this here by adding the 'if status:' check, which reproduces
the original behavior of just ignoring other job statuses.